chore: align type annotations with modern PEP 585 and PEP 604 guidelines

+++ src/cxas_scrapi/cli/trace_cli.py
@@ -460,7 +460,7 @@ def _bar_block(
     label_w = min(40, max(len(str(k)) for k, _ in items))
     out: list[str] = ["```"]
     for label, n in items:
-        bar = "█" * max(1, int(round(width * n / peak)))
+        bar = "█" * max(1, round(width * n / peak))
         pct = ""
         if total:
             pct = f"  ({n / total:.0%})"
